move router keys to separate file
